Made manager.cpp compatible with lab machines
They are running older versions, needed static casts

+++ Lab1Template/source/manager.cpp
@@ -15,7 +15,7 @@ Manager::Manager(string newName, Camera cam)
 void Manager::reportStats()
 {
 	cout << name + "'s Stats:" << endl;
-	cout << "Gold: " + to_string(gold) << endl;
-	cout << "Beer: " + to_string(beer) << endl;
-	cout << "Mercenarys: " + to_string(mercs) << endl;
+	cout << "Gold: " + to_string(static_cast<long double>(gold)) << endl;
+	cout << "Beer: " + to_string(static_cast<long long int>(beer)) << endl;
+	cout << "Mercenarys: " + to_string(static_cast<long long int>(mercs)) << endl;
 }
